ALS-2476 Remove test_var from all codepipeline builds

+++ scripts/terraform-local-destroy.sh
@@ -2,7 +2,17 @@
 
 set -e
 
-rm -rf terraform/env_configs
-git clone https://github.com/ministryofjustice/hmpps-env-configs.git terraform/env_configs
-cd ${HMPPS_BUILD_WORK_DIR}/components
+if [[ "$ci_components_flag" = true ]] ; then
+    source ${HMPPS_BUILD_WORK_DIR}/ci_env_configs/dev.properties
+    cd ${HMPPS_BUILD_WORK_DIR}/ci-components
+else
+      rm -rf terraform/env_configs
+      git clone https://github.com/ministryofjustice/hmpps-env-configs.git terraform/env_configs
+
+      source ${HMPPS_BUILD_WORK_DIR}/env_configs/${environment_name}/${environment_name}.properties
+
+      cd ${HMPPS_BUILD_WORK_DIR}/components
+fi
+
+
 /home/tools/data/scripts/stack-destroy.sh
